Remove panics from across the application
Most of the instances of panics have been removed, either by logging the error
and allowing the next time it is run to have fixed the issue, or by returning
an error up the callstack.

Create a UI Handler that handles the lifecycle of the UI server, starting it
and stopping as needed, since it was blocking the app before.