Remove second historical sync process
By switching the order of the chain head sync
and the historical sync, we can remove the need
to listen for the first chain head event to
the second stage historical sync.

By listening to the chain head events first, we
can be sure we haven't missed any block headers
before starting the historical sync.

Instead, we may get duplicate blocks processed,
but the application already handles duplicate
data.