Combine RPC client methods with GraphQL types
The ethereum client is able to execute both RPC
calls and GraphQL queries.

Users of the application should not be privy to
the internals of how a call is made, which is
true for RPC calls where the required methods
exist within the `client/utils.go` file.

This brings in the GraphQL methods to be the
same, freeing other areas of the appliaction
from needing to know how to use the client
directly.