Store storage entries and k/v pairs in a list
ElasticSearch has a (default) limit on the number
of properties a document can have as 1000. A
storage entry may go well beyond that, so instead
they can be stored as k/v pairs, so that the
number of properties is fixed at 2.

This only applies to ElasticSearch. The rest of
the application will see the full map as before.