Use workers for storage filtering concurrency
The storage filter fires off all storage
requests at the same time, which can quickly
overload both the reporting process and the
attached Quorum process, resulting in a lot of
memory usage and request timeouts.

To reduce the concurrency strain, this commit
introduces workers for pulling storage and
saving the storage. A pre-set number of workers
will fetch the storage from Quorum and then pass
it onto a preset number of workers who will save
that state in the database.